Clean up based on review
- Renamed `spectrum_wrappers` to just `spectrum`
- Externalize logging
  - Requires a change in web-client-ui as well
- Other cleanups
