+++ plugins/ui/src/deephaven/ui/__init__.py
@@ -9,7 +9,7 @@ from .components import *
 from .hooks import *
 from .object_types import *
 
-__version__ = "0.0.1.dev0"
+__version__ = "0.0.1.dev1"
 
 
 class UIRegistration(Registration):
