+++ plugins/ui/src/deephaven/ui/renderer/Renderer.py
@@ -1,6 +1,6 @@
 import logging
 from typing import Any
-from .._internal import RenderContext, get_component_qualname
+from .._internal import RenderContext
 from ..elements import Element
 from .RenderedNode import RenderedNode
