+++ plugins/ui/src/js/src/SpectrumElementUtils.ts
@@ -13,7 +13,7 @@ import {
   View,
 } from '@adobe/react-spectrum';
 import { ValueOf } from '@deephaven/utils';
-import { ActionButton, Slider, TextField } from './spectrum_wrappers';
+import { ActionButton, Slider, TextField } from './spectrum';
 import { ELEMENT_KEY, ElementNode, isElementNode } from './ElementUtils';
 
 export const SPECTRUM_ELEMENT_TYPE_PREFIX = 'deephaven.ui.spectrum.';
