+++ plugins/ui/src/js/src/spectrum/ActionButton.tsx
@@ -11,6 +11,8 @@ function ActionButton(
 
   const onPress = useCallback(
     e => {
+      // The PressEvent from React Spectrum is not serializable (contains circular references). We're just dropping the event here but we should probably convert it.
+      // TODO(#76): Need to serialize PressEvent and send with the callback instead of just dropping it.
       propOnPress?.();
     },
     [propOnPress]
