+++ plugins/ui/src/js/src/spectrum/index.ts
@@ -2,6 +2,7 @@
  * Wrappers for Spectrum components.
  * Used if we want to mediate some of the props values passed to the Spectrum components.
  * For example, we may want to debounce sending the change for a text field, and also keep the value on the client side until the change is sent.
+ * Or in the case of event handlers, we may want to wrap the event handler to serialize the event correctly.
  */
 export { default as ActionButton } from './ActionButton';
 export { default as Slider } from './Slider';
